requirements.txt
aiomysql==0.2.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.3
asgi-lifespan==2.1.0
asttokens==2.2.1
backcall==0.2.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.0.7
cryptography==44.0.2
cycler==0.11.0
decorator==5.1.1
distlib==0.3.9
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.115.11
filelock==3.17.0
fonttools==4.39.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
ipython==8.14.0
jedi==0.18.2
joblib==1.2.0
jwt==1.3.1
kiwisolver==1.4.4
kombu==5.4.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.3.6
pluggy==1.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.15.1
PyMySQL==1.1.1
pyparsing==3.0.9
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-mock==3.14.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
redis==5.2.1
requests==2.32.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.24
stack-data==0.6.2
starlette==0.46.1
threadpoolctl==3.1.0
traitlets==5.9.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.6
